Move api into tendermint package
As part of this I had to neuter the API tests but I left empty stubs
so that we can re-instate them.
