+++ consensus/tendermint/api_test.go
@@ -1,0 +1,27 @@
+package tendermint
+
+import (
+	"testing"
+)
+
+// TODO re-implement these tests
+
+func TestGetCommittee(t *testing.T) {
+}
+
+func TestGetCommitteeAtHash(t *testing.T) {
+	t.Run("unknown block given, error returned", func(t *testing.T) {
+	})
+
+	t.Run("valid block given, committee returned", func(t *testing.T) {
+	})
+}
+
+func TestAPIGetContractABI(t *testing.T) {
+}
+
+func TestAPIGetContractAddress(t *testing.T) {
+}
+
+func TestAPIGetWhitelist(t *testing.T) {
+}
