+++ consensus/tendermint/core_backend.go
@@ -6,6 +6,7 @@ import (
 	"github.com/clearmatics/autonity/consensus"
 	"github.com/clearmatics/autonity/core"
 	"github.com/clearmatics/autonity/core/types"
+	"github.com/clearmatics/autonity/rpc"
 )
 
 type Tendermint interface {
@@ -13,4 +14,5 @@ type Tendermint interface {
 	Start(ctx context.Context, blockchain *core.BlockChain)
 	Stop()
 	Seal(chain consensus.ChainReader, block *types.Block, results chan<- *types.Block, stop <-chan struct{}) error
+	APIs(chain consensus.ChainReader) []rpc.API
 }
