+++ internal/ethapi/api.go
@@ -35,7 +35,6 @@ import (
 	"github.com/clearmatics/autonity/common/hexutil"
 	"github.com/clearmatics/autonity/common/math"
 	"github.com/clearmatics/autonity/consensus/ethash"
-	"github.com/clearmatics/autonity/contracts/autonity"
 	"github.com/clearmatics/autonity/core"
 	"github.com/clearmatics/autonity/core/rawdb"
 	"github.com/clearmatics/autonity/core/types"
@@ -523,10 +522,6 @@ func (s *PublicBlockChainAPI) ChainId() *hexutil.Big {
 func (s *PublicBlockChainAPI) BlockNumber() hexutil.Uint64 {
 	header, _ := s.b.HeaderByNumber(context.Background(), rpc.LatestBlockNumber) // latest header should always be available
 	return hexutil.Uint64(header.Number.Uint64())
-}
-
-func (s *PublicBlockChainAPI) AutonityContract() *autonity.Contract {
-	return s.b.AutonityContract()
 }
 
 // GetBalance returns the amount of wei for the given address in the state of the
